// src/api/model/flagDetailResult.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */
import type { FlagHostResult } from './flagHostResult';
import type { FlagScheduleResult } from './flagScheduleResult';
import type { ParentFlagResult } from './parentFlagResult';
import type { ParticipantResult } from './participantResult';

export interface FlagDetailResult {
  id?: number;
  title?: string;
  description?: string;
  capacity?: number;
  participantCount?: number;
  parentFlagId?: number;
  status?: string;
  schedule?: FlagScheduleResult;
  host?: FlagHostResult;
  parentFlag?: ParentFlagResult;
  participants?: ParticipantResult[];
  isHost?: boolean;
}