// src/api/model/memorialResult.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */

export interface MemorialResult {
  id?: number;
  writerId?: number;
  nickname?: string;
  profileImageUrl?: string;
  content?: string;
  createdAt?: string;
}